improve MVC structure #78
